+++ packages/backend/src/ee/index.ts
@@ -21,6 +21,7 @@ import { CommercialCatalogService } from './services/CommercialCatalogService';
 import { CommercialSlackIntegrationService } from './services/CommercialSlackIntegrationService';
 import { EmbedService } from './services/EmbedService/EmbedService';
 import { ScimService } from './services/ScimService/ScimService';
+import { SupportService } from './services/SupportService/SupportService';
 
 type EnterpriseAppArguments = Pick<
     AppArguments,
@@ -116,6 +117,19 @@ export async function getEnterpriseAppArguments(): Promise<EnterpriseAppArgument
                     slackAuthenticationModel:
                         models.getSlackAuthenticationModel() as CommercialSlackAuthenticationModel,
                     analytics: context.lightdashAnalytics,
+                }),
+            supportService: ({ models, context, repository, clients }) =>
+                new SupportService({
+                    analytics: context.lightdashAnalytics,
+                    projectModel: models.getProjectModel(),
+                    savedChartModel: models.getSavedChartModel(),
+                    dashboardModel: models.getDashboardModel(),
+                    spaceModel: models.getSpaceModel(),
+                    s3Client: clients.getS3Client(),
+                    organizationModel: models.getOrganizationModel(),
+                    unfurlService: repository.getUnfurlService(),
+                    projectService: repository.getProjectService(),
+                    lightdashConfig: context.lightdashConfig,
                 }),
         },
         modelProviders: {