+++ packages/frontend/src/components/NavBar/HelpMenu.tsx
@@ -1,21 +1,25 @@
 import { LightdashMode } from '@lightdash/common';
 import { Button, Menu } from '@mantine/core';
+import { modals } from '@mantine/modals';
 import {
     IconBook,
     IconHelp,
     IconMessageCircle2,
     IconMessages,
+    IconSos,
     IconUsers,
 } from '@tabler/icons-react';
 import { type FC } from 'react';
 import { useIntercom } from 'react-use-intercom';
 import useHealth from '../../hooks/health/useHealth';
+import SupportDrawerContent from '../../providers/SupportDrawer/SupportDrawerContent';
 import LargeMenuItem from '../common/LargeMenuItem';
 import MantineIcon from '../common/MantineIcon';
 
 const HelpMenu: FC = () => {
     const health = useHealth();
     const isCloudCustomer = health.data?.mode === LightdashMode.CLOUD_BETA;
+    const isDevelopment = health.data?.mode === LightdashMode.DEV;
 
     const { show: showIntercom } = useIntercom();
 
@@ -77,6 +81,24 @@ const HelpMenu: FC = () => {
                     description="Submit a feature request or bug report to improve Lightdash."
                     icon={IconMessageCircle2}
                 />
+                {(isCloudCustomer || isDevelopment) && (
+                    <LargeMenuItem
+                        component="a"
+                        onClick={() => {
+                            modals.open({
+                                id: 'support-drawer',
+                                title: 'Share with Lightdash Support',
+                                size: 'lg',
+                                children: <SupportDrawerContent />,
+                                yOffset: 100,
+                                zIndex: 1000,
+                            });
+                        }}
+                        title="Report an issue to Lightdash Support"
+                        description="Share a detailed issue report with Lightdash Support"
+                        icon={IconSos}
+                    />
+                )}
             </Menu.Dropdown>
         </Menu>
     );